@@ -8,10 +8,13 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "@fortawesome/vue-fontawesome": "^3.0.6",
+    "axios": "^1.6.8",
     "core-js": "^3.8.3",
     "flowbite": "^2.3.0",
-    "vue": "^3.2.13",
-    "vue-router": "^4.0.3"
+    "vue": "^3.4.26",
+    "vue-router": "^4.0.3",
+    "vue-toastification": "^2.0.0-rc.5"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
