@@ -2,5 +2,12 @@ import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
 import './index.css'
+import Toast from 'vue-toastification'
+import 'vue-toastification/dist/index.css'
 
-createApp(App).use(router).mount('#app')
+const app = createApp(App)
+
+app.use(Toast)
+app.use(router)
+
+app.mount('#app')
